refactor: PaymentType 패키지 변경
