+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/domain/Character.java
@@ -1,7 +1,6 @@
 package com.startingblue.fourtooncookie.character.domain;
 
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
-import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import jakarta.persistence.*;
 import jakarta.validation.*;
 import jakarta.validation.constraints.NotBlank;
