+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/domain/PaymentType.java
@@ -1,0 +1,5 @@
+package com.startingblue.fourtooncookie.character.domain;
+
+public enum PaymentType {
+    FREE, PAID
+}
