--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/global/domain/PaymentType.java
@@ -1,5 +1,0 @@
-package com.startingblue.fourtooncookie.global.domain;
-
-public enum PaymentType {
-    FREE, PAID
-}
