+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/character/domain/CharacterRepositoryTest.java
@@ -3,7 +3,6 @@ package com.startingblue.fourtooncookie.character.domain;
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
 import com.startingblue.fourtooncookie.artwork.ArtworkRepository;
 import com.startingblue.fourtooncookie.character.CharacterRepository;
-import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
