+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/character/domain/CharacterTest.java
@@ -1,7 +1,6 @@
 package com.startingblue.fourtooncookie.character.domain;
 
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
-import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
